[#2384] Refactor model managers
Refactor the model managers for Employment, Organisation and Project using the
Manager.from_queryset() method

Add OrganisationModelTestCase.test_chaining_managers() to "prove" the
chainability of querysets created in the new way

Add OrganisationModelTestCase.test_order_queryset_by_lower_name() to make sure
Organisation querysets are ordered by the org name by default

Also a couple of flake8 fixes
